rename cwSfExecutionEventToDb lambda to cwSfEventToDbRecords
