+++ packages/api/tests/lambdas/fixtures/cw-sf-event-to-db-records/execution-running-event.json
